Refactor, remove old components and make code more readable
