+++ src/views/notificationprofile/NotificationProfileView.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import "./NotificationProfileView.css";
 import { withRouter } from "react-router-dom";
 import { Helmet } from "react-helmet";
-import { NotificationProfileList } from "../../components/profilelist/ProfileList";
+import NotificationProfileList from "../../components/notificationprofile/NotificationProfileList";
 import { Container } from "@material-ui/core";
 
 const NotificationProfileView: React.FC = () => {
